requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.5.0=py39haa95532_0
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py39haa95532_0
babel=2.11.0=py39haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py39haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py39h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=py39haa95532_0
comm=0.1.2=py39haa95532_0
cryptography=39.0.1=py39h21b164f_0
cuda=11.6.1=0
cuda-cccl=11.6.55=0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=0
cuda-cudart-dev=11.6.55=0
cuda-cuobjdump=11.6.124=0
cuda-cupti=11.6.124=0
cuda-cuxxfilt=11.6.124=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=12.0.0=0
cuda-nvcc=11.6.124=0
cuda-nvdisasm=12.0.76=0
cuda-nvml-dev=11.6.55=0
cuda-nvprof=12.0.90=0
cuda-nvprune=11.6.124=0
cuda-nvrtc=11.6.124=0
cuda-nvrtc-dev=11.6.124=0
cuda-nvtx=11.6.124=0
cuda-nvvp=12.0.90=0
cuda-runtime=11.6.1=0
cuda-sanitizer-api=12.0.90=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
cudatoolkit=11.8.0=h09e9e62_11
cvxpy=1.3.1=py39hcbf5309_0
cvxpy-base=1.3.1=py39h2ba5b7c_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dppy=0.3.2=pypi_0
ecos=2.0.12=py39h080aedc_0
eigen=3.3.7=h59b6b97_1
entrypoints=0.4=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
faiss=1.7.2=py39cuda112h50edf10_3_cuda
faiss-gpu=1.7.2=h949689a_3
ffmpeg=4.2.2=he774522_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
git=2.34.1=haa95532_0
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h5py=3.7.0=py39h3de5c98_0
hdf5=1.10.6=h1756f20_1
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py39haa95532_0
importlib-metadata=6.0.0=py39haa95532_0
importlib_metadata=6.0.0=hd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.19.2=py39hd4e2768_0
ipython=8.12.0=py39haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.4=py39haa95532_0
jedi=0.18.1=py39haa95532_1
jinja2=3.1.2=py39haa95532_0
joblib=1.1.1=py39haa95532_0
jpeg=9e=h2bbff1b_1
json5=0.9.6=pyhd3eb1b0_0
jsonpatch=1.32=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=py39haa95532_0
jupyter=1.0.0=py39haa95532_8
jupyter_client=8.1.0=py39haa95532_0
jupyter_console=6.6.3=py39haa95532_0
jupyter_core=5.3.0=py39haa95532_0
jupyter_server=1.23.4=py39haa95532_0
jupyterlab=3.5.3=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.22.0=py39haa95532_0
jupyterlab_widgets=3.0.5=py39haa95532_0
kiwisolver=1.4.4=py39hd77b12b_0
krb5=1.19.4=h5b6d351_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcublas=11.9.2.110=0
libcublas-dev=11.9.2.110=0
libcufft=10.7.1.112=0
libcufft-dev=10.7.1.112=0
libcurand=10.3.1.50=0
libcurand-dev=10.3.1.50=0
libcusolver=11.3.4.124=0
libcusolver-dev=11.3.4.124=0
libcusparse=11.7.2.124=0
libcusparse-dev=11.7.2.124=0
libdeflate=1.17=h2bbff1b_0
libfaiss=1.7.2=cuda112h33bf9e0_3_cuda
libfaiss-avx2=1.7.2=cuda112h1234567_3_cuda
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
liblapack=3.9.0=5_hd5c7e75_netlib
libnpp=11.6.3.124=0
libnpp-dev=11.6.3.124=0
libnvjpeg=11.6.2.124=0
libnvjpeg-dev=11.6.2.124=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libqdldl=0.1.5=h63175ca_1
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
lxml=4.9.2=py39h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py39h2bbff1b_0
matplotlib=3.5.3=py39hcbf5309_2
matplotlib-base=3.5.3=py39hd77b12b_0
matplotlib-inline=0.1.6=py39haa95532_0
mistune=0.8.4=py39h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nbclassic=0.5.5=py39haa95532_0
nbclient=0.5.13=py39haa95532_0
nbconvert=6.5.4=py39haa95532_0
nbformat=5.7.0=py39haa95532_0
nest-asyncio=1.5.6=py39haa95532_0
networkx=3.0=pypi_0
notebook=6.5.4=py39haa95532_0
notebook-shim=0.2.2=py39haa95532_0
nsight-compute=2022.4.0.15=0
numexpr=2.8.4=py39h5b0cc5e_0
numpy=1.22.4=pypi_0
numpy-base=1.23.5=py39h4da318b_0
opencv=4.6.0=py39hf11a4ad_3
opencv-contrib-python=4.7.0.72=pypi_0
openssl=1.1.1t=h2bbff1b_0
osqp=0.6.2.post0=py39h2e25243_3
packaging=23.0=py39haa95532_0
pandas=1.5.3=py39hf11a4ad_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py39hd77b12b_0
pip=23.0.1=py39haa95532_0
platformdirs=2.5.2=py39haa95532_0
ply=3.11=py39haa95532_0
pooch=1.4.0=pyhd3eb1b0_0
prometheus_client=0.14.1=py39haa95532_0
prompt-toolkit=3.0.36=py39haa95532_0
prompt_toolkit=3.0.36=hd3eb1b0_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py39haa95532_0
pyopenssl=23.0.0=py39haa95532_0
pyparsing=3.0.9=py39haa95532_0
pyqt=5.15.7=py39hd77b12b_0
pyqt5-sip=12.11.0=py39hd77b12b_0
pyrsistent=0.18.0=py39h196d8e1_0
pysocks=1.7.1=py39haa95532_0
python=3.9.16=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39haa95532_0
python_abi=3.9=2_cp39
pytorch=1.13.1=py3.9_cuda11.6_cudnn8_0
pytorch-cuda=11.6=h867d48c_1
pytorch-metric-learning=2.0.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.7=py39haa95532_0
pywin32=305=py39h2bbff1b_0
pywinpty=2.0.10=py39h5da7b33_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=py39hd77b12b_0
qdldl-python=0.1.5.post2=py39h2ba5b7c_0
qpth=0.0.15=pypi_0
qt-main=5.15.2=he8e5bd7_8
qt-webengine=5.15.9=hb9a9bb5_5
qtconsole=5.4.2=py39haa95532_0
qtpy=2.2.0=py39haa95532_0
qtwebkit=5.212=h2bbfb41_5
requests=2.29.0=py39haa95532_0
scikit-learn=1.0.2=py39hf11a4ad_1
scipy=1.10.1=py39h321e85e_0
scs=3.2.3=py39h212b5b4_0
seaborn=0.11.2=pyhd3eb1b0_0
sip=6.6.2=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=66.0.0=py39haa95532_0
sniffio=1.2.0=py39haa95532_1
soupsieve=2.4=py39haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.17.1=py39haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py39haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py39haa95532_0
torchaudio=0.13.1=pypi_0
torchnet=0.0.4=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=py39h2bbff1b_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.7.1=py39haa95532_0
typing-extensions=4.5.0=py39haa95532_0
typing_extensions=4.5.0=py39haa95532_0
tzdata=2023c=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
urllib3=1.26.15=py39haa95532_0
vc=14.2=h21ff451_1
visdom=0.2.4=pypi_0
vs2015_runtime=14.34.31931=h4c5c07a_11
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
websocket-client=1.5.1=pypi_0
wheel=0.38.4=py39haa95532_0
widgetsnbextension=4.0.5=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
winpty=0.4.3=4
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=he774522_0
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py39haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
